refactor(assistant): shrink orchestrator/tools functions and split worktree module
