+++ src/quodeq/assistant/orchestrator.py
@@ -92,6 +92,68 @@ def _mcp_server_args(request: TurnRequest, tool_ctx: ToolContext) -> list[str]:
     return args
 
 
+def _resolve_write_grant(request: TurnRequest, repository: AssistantStore,
+                          tool_ctx: ToolContext) -> tuple[bool, ToolContext]:
+    """Server-derived write grant, mirror of web_tools_on: the client flag
+    alone is never enough. Requires an attached LOCAL git repo and a
+    provider whose tool wiring is per-invocation isolated. When granted,
+    ensures the session worktree exists and returns a tool_ctx pointed at
+    it."""
+    write_on = (request.write_enabled and not tool_ctx.read_only
+                and tool_ctx.repo_root is not None
+                and (tool_ctx.repo_root / ".git").exists()
+                and write_safe_provider(request.provider))
+    if write_on:
+        manager = ensure_session_worktree(
+            repository, repo_root=tool_ctx.repo_root,
+            project_id=tool_ctx.project_id, session_id=request.session_id)
+        tool_ctx = replace(tool_ctx, worktree_dir=manager.path)
+    return write_on, tool_ctx
+
+
+def _run_cli_engine(request: TurnRequest, tool_ctx: ToolContext, messages: list[dict],
+                     skill, repository: AssistantStore, emit, cancel, cli_turn_fn) -> str:
+    skill_block = (f"[skill:{skill.name}]\n{skill.instructions}"
+                   if skill is not None else "")
+    return cli_turn_fn(
+        messages=messages,
+        config=CliTurnConfig(
+            provider=request.provider, model=request.model,
+            scratch_base=tool_ctx.repository.db_path.parent,
+            mcp_server_args=_mcp_server_args(request, tool_ctx),
+            db_path=tool_ctx.repository.db_path,
+            web_enabled=request.web_enabled,
+            system_prompt=messages[0]["content"],
+            skill_block=skill_block,
+            worktree_dir=tool_ctx.worktree_dir,
+        ),
+        session_id=request.session_id,
+        prior_session_id=(repository.get_session(request.session_id) or {}).get("cli_session_id"),
+        repository=repository, emit=emit, cancel=cancel,
+    )
+
+
+def _run_api_engine(request: TurnRequest, tool_ctx: ToolContext, messages: list[dict],
+                     skill, web_tools_on: bool, write_on: bool, emit, cancel,
+                     turn_fn, capability_fn) -> str:
+    config = ApiTurnConfig(
+        api_base=request.api_base, api_key=request.api_key,
+        model=request.model,
+        native_tools=capability_fn(request.provider, request.api_base,
+                                   request.model),
+        max_tool_iterations=max(
+            SKILL_MAX_TOOL_ITERATIONS if skill is not None else MAX_TOOL_ITERATIONS,
+            WRITE_MAX_TOOL_ITERATIONS if write_on else 0),
+    )
+    registry = build_registry(tool_ctx)
+    if web_tools_on:
+        register_web_tools(registry)
+    if write_on:
+        register_write_tools(registry, tool_ctx)
+    return turn_fn(messages=messages, config=config,
+                   registry=registry, emit=emit, cancel=cancel)
+
+
 def run_turn(request: TurnRequest, *, repository: AssistantStore,
              tool_ctx: ToolContext, turn_fn=None, capability_fn=None,
              cli_turn_fn=None, cancel: CancelToken | None = None) -> None:
@@ -114,59 +176,16 @@ def run_turn(request: TurnRequest, *, repository: AssistantStore,
         # In-process web tools are local-API-only: claude gets NATIVE web
         # tools via argv, and cloud APIs (openrouter/custom) stay excluded.
         web_tools_on = request.web_enabled and request.provider in LOCAL_PROVIDERS
-        # Server-derived write grant, mirror of web_tools_on: the client flag
-        # alone is never enough. Requires an attached LOCAL git repo and a
-        # provider whose tool wiring is per-invocation isolated.
-        write_on = (request.write_enabled and not tool_ctx.read_only
-                    and tool_ctx.repo_root is not None
-                    and (tool_ctx.repo_root / ".git").exists()
-                    and write_safe_provider(request.provider))
-        if write_on:
-            manager = ensure_session_worktree(
-                repository, repo_root=tool_ctx.repo_root,
-                project_id=tool_ctx.project_id, session_id=request.session_id)
-            tool_ctx = replace(tool_ctx, worktree_dir=manager.path)
+        write_on, tool_ctx = _resolve_write_grant(request, repository, tool_ctx)
         messages = [{"role": "system",
                      "content": build_system_prompt(skill=skill,
                                                     web_enabled=web_tools_on,
                                                     write_enabled=write_on)},
                     *({"role": m["role"], "content": m["content"]} for m in history)]
         if _provider_type(request.provider) == "cli":
-            skill_block = (f"[skill:{skill.name}]\n{skill.instructions}"
-                           if skill is not None else "")
-            final = cli_turn_fn(
-                messages=messages,
-                config=CliTurnConfig(
-                    provider=request.provider, model=request.model,
-                    scratch_base=tool_ctx.repository.db_path.parent,
-                    mcp_server_args=_mcp_server_args(request, tool_ctx),
-                    db_path=tool_ctx.repository.db_path,
-                    web_enabled=request.web_enabled,
-                    system_prompt=messages[0]["content"],
-                    skill_block=skill_block,
-                    worktree_dir=tool_ctx.worktree_dir,
-                ),
-                session_id=request.session_id,
-                prior_session_id=(repository.get_session(request.session_id) or {}).get("cli_session_id"),
-                repository=repository, emit=emit, cancel=cancel,
-            )
+            final = _run_cli_engine(request, tool_ctx, messages, skill, repository, emit, cancel, cli_turn_fn)
         else:
-            config = ApiTurnConfig(
-                api_base=request.api_base, api_key=request.api_key,
-                model=request.model,
-                native_tools=capability_fn(request.provider, request.api_base,
-                                           request.model),
-                max_tool_iterations=max(
-                    SKILL_MAX_TOOL_ITERATIONS if skill is not None else MAX_TOOL_ITERATIONS,
-                    WRITE_MAX_TOOL_ITERATIONS if write_on else 0),
-            )
-            registry = build_registry(tool_ctx)
-            if web_tools_on:
-                register_web_tools(registry)
-            if write_on:
-                register_write_tools(registry, tool_ctx)
-            final = turn_fn(messages=messages, config=config,
-                            registry=registry, emit=emit, cancel=cancel)
+            final = _run_api_engine(request, tool_ctx, messages, skill, web_tools_on, write_on, emit, cancel, turn_fn, capability_fn)
         repository.add_message(request.session_id, "assistant", final)
         emit({"type": "done"})
     except TurnCancelled as exc:
